商家管理 > 商家录入商品完成

+++ pyg_dao/src/main/java/com/songxin/pyg/mapper/BrandMapper.java
@@ -35,7 +35,7 @@ public interface BrandMapper {
      * @author fishsx
      * @date 2018/12/4 下午8:23
      */
-    TbBrand findOneById(Integer id);
+    TbBrand findOneById(Long id);
 
     /**
      * 修改品牌
